Attempt to read preferences and data directory from `the-application` package
... to enable integrating this package into ActivityLog2 (and other
application), the code will look for a package named `the-application` and
attempt to load it to obtain the `get-pref`, `put-pref` and `data-directory`
files.  This way, the application can configure what preference files are
used and where should the cached map tiles be placed.

This has no effect to casual users of the package -- the Racket preferences
are used by default and the racket configuration file path is used for the
cached map tiles (just as before)