Don't delay a `resize-to-fit` for a sub-group (workaround)
When calling the `resize-to-fit` inside an edit sequence, the request is
delayed.  Unfortunately, when the edit sequence is closed, a `resize-to-fit`
is done for the entire map, regardless of the `group` argument of the original
`resize-to-fit` call.

As a workaround, if `resize-to-fit` is called with a `group` argument, the
operation is never delayed, even if the call is inside an edit sequence.

Also updated build to use 8.0-cs