Moved point-cloud processing into a separate OS thread
Use a separate `racket/place` for processing `geoids` in `gbuffers` and
constructing draw buffers.  A few things had to change due to this, most
notable is that the draw buffer is now a `flvector` so it can be shared
between the place and the main thread without copying it.

`geoid` lists are still sent through the place channel which may undo some of
the performance benefits, but the GUI responsiveness is better.